- refactoring to more explicitly use macroscopic network and macroscopic link segments where appropriate - preparing for better support for removing id gaps through network builder implementations

+++ src/main/java/org/planit/network/physical/macroscopic/MacroscopicLinkSegmentTypeImpl.java
@@ -57,7 +57,7 @@ public class MacroscopicLinkSegmentTypeImpl implements MacroscopicLinkSegmentTyp
    * @param groupId contiguous id generation within this group for instances of this class
    * @return id of this link segment
    */
-  protected static int generateMacroscopicLinkSegmentTypeId(final IdGroupingToken groupId) {
+  protected static long generateMacroscopicLinkSegmentTypeId(final IdGroupingToken groupId) {
     return IdGenerator.generateId(groupId, MacroscopicLinkSegmentType.class);
   }
